+++ lib/array/limit/method.ts
@@ -1,0 +1,3 @@
+const method = (arr: any[], limit: number) => arr.slice(0, limit);
+
+export = method;