+++ lib/array/limit/test.ts
@@ -1,0 +1,7 @@
+import "./index";
+
+describe("Array.prototype.limit", () => {
+  test("[2, 1, 2, 5].limit(2) returns [2,1]", () => {
+    expect([2, 1, 2, 5].limit(2)).toEqual([2, 1]);
+  });
+});