+++ lib/array/skip/test.ts
@@ -1,0 +1,7 @@
+import "./index";
+
+describe("Array.prototype.skip", () => {
+  test("[2, 1, 2, 5].skip(1) returns [1,2,5]", () => {
+    expect([2, 1, 2, 5].skip(1)).toEqual([1, 2, 5]);
+  });
+});